Add files via upload
First I changed the original data 4hours.csv into our used standard data by adding 4 bits of time of day and 2 bits of day of week. Then divided the standard data into two different parts, one is used for training, the other is used for test. Dividing the data set into two parts is to ensure  the training data and the test data have the similar mean and variance, that is, have the same distribution.